# Conflicts:
#	android/AndroidManifest.xml
#	android/testing/AndroidManifest.xml
#	src/Airspace/NearestAirspace.cpp
#	src/Android/Main.cpp
#	src/event/net/cares/Init.cxx

+++ src/Dialogs/Plane/PlaneListDialog.cpp
@@ -22,6 +22,7 @@
 #include "Interface.hpp"
 #include "Language/Language.hpp"
 #include "util/StringAPI.hxx"
+#include "util/StringCompare.hxx"
 
 #include <vector>
 #include <cassert>
@@ -38,10 +39,10 @@ class PlaneListWidget final
     StaticString<32> name;
     AllocatedPath path;
 
-    ListItem(const TCHAR *_name, Path _path)
+    ListItem(tstring_view _name, Path _path) noexcept
       :name(_name), path(_path) {}
 
-    bool operator<(const ListItem &i2) const {
+    bool operator<(const ListItem &i2) const noexcept {
       return StringCollate(name, i2.name) < 0;
     }
   };
@@ -51,32 +52,35 @@ class PlaneListWidget final
     std::vector<ListItem> &list;
 
   public:
-    PlaneFileVisitor(std::vector<ListItem> &_list):list(_list) {}
+    PlaneFileVisitor(std::vector<ListItem> &_list) noexcept:list(_list) {}
 
     void Visit(Path path, Path filename) override {
-      list.emplace_back(filename.c_str(), path);
+      tstring_view name{filename.c_str()};
+      RemoveSuffix(name, tstring_view{_T(".xcp")});
+
+      list.emplace_back(name, path);
     }
   };
 
   WndForm *form;
-  Button *edit_button, *delete_button, *load_button;
+  Button *edit_button, *copy_button, *delete_button, *load_button;
 
   std::vector<ListItem> list;
 
   TwoTextRowsRenderer row_renderer;
 
 public:
-  void CreateButtons(WidgetDialog &dialog);
+  void CreateButtons(WidgetDialog &dialog) noexcept;
 
 private:
-  void UpdateList();
-  bool Load(unsigned i);
-  bool LoadWithDialog(unsigned i);
+  void UpdateList() noexcept;
+  bool Load(unsigned i) noexcept;
+  bool LoadWithDialog(unsigned i) noexcept;
 
-  void LoadClicked();
-  void NewClicked();
-  void EditClicked();
-  void DeleteClicked();
+  void LoadClicked() noexcept;
+  void NewClicked() noexcept;
+  void EditClicked(bool copy) noexcept;
+  void DeleteClicked() noexcept;
 
 public:
   /* virtual methods from class Widget */
@@ -96,7 +100,7 @@ protected:
 };
 
 void
-PlaneListWidget::UpdateList()
+PlaneListWidget::UpdateList() noexcept
 {
   list.clear();
 
@@ -119,12 +123,13 @@ PlaneListWidget::UpdateList()
 }
 
 void
-PlaneListWidget::CreateButtons(WidgetDialog &dialog)
+PlaneListWidget::CreateButtons(WidgetDialog &dialog) noexcept
 {
   form = &dialog;
 
   dialog.AddButton(_("New"), [this](){ NewClicked(); });
-  edit_button = dialog.AddButton(_("Edit"), [this](){ EditClicked(); });
+  edit_button = dialog.AddButton(_("Edit"), [this](){ EditClicked(false); });
+  copy_button = dialog.AddButton(_("Copy"), [this](){ EditClicked(true); });
   delete_button = dialog.AddButton(_("Delete"), [this](){ DeleteClicked(); });
   load_button = dialog.AddButton(_("Activate"), [this](){ LoadClicked(); });
 }
@@ -152,11 +157,15 @@ PlaneListWidget::OnPaintItem(Canvas &canvas, const PixelRect rc,
   } else
     row_renderer.DrawFirstRow(canvas, rc, list[i].name);
 
-  row_renderer.DrawSecondRow(canvas, rc, list[i].path.c_str());
+  Path path = list[i].path;
+  if (auto relative_path = RelativePath(path); relative_path != nullptr)
+    path = relative_path;
+
+  row_renderer.DrawSecondRow(canvas, rc, path.c_str());
 }
 
 static bool
-LoadFile(Path path)
+LoadFile(Path path) noexcept
 {
   ComputerSettings &settings = CommonInterface::SetComputerSettings();
 
@@ -173,7 +182,7 @@ LoadFile(Path path)
 }
 
 bool
-PlaneListWidget::Load(unsigned i)
+PlaneListWidget::Load(unsigned i) noexcept
 {
   assert(i < list.size());
 
@@ -181,7 +190,7 @@ PlaneListWidget::Load(unsigned i)
 }
 
 bool
-PlaneListWidget::LoadWithDialog(unsigned i)
+PlaneListWidget::LoadWithDialog(unsigned i) noexcept
 {
   bool result = Load(i);
   if (!result) {
@@ -196,16 +205,16 @@ PlaneListWidget::LoadWithDialog(unsigned i)
 }
 
 inline void
-PlaneListWidget::LoadClicked()
+PlaneListWidget::LoadClicked() noexcept
 {
   if (LoadWithDialog(GetList().GetCursorIndex()))
     form->SetModalResult(mrOK);
 }
 
 inline void
-PlaneListWidget::NewClicked()
+PlaneListWidget::NewClicked() noexcept
 {
-  Plane plane = CommonInterface::GetComputerSettings().plane;
+  Plane plane{};
 
   while (dlgPlaneDetailsShowModal(plane)) {
     if (plane.registration.empty()) {
@@ -240,8 +249,8 @@ PlaneListWidget::NewClicked()
   }
 }
 
-inline void
-PlaneListWidget::EditClicked()
+void
+PlaneListWidget::EditClicked(bool copy) noexcept
 {
   assert(GetList().GetCursorIndex() < list.size());
 
@@ -262,7 +271,7 @@ PlaneListWidget::EditClicked()
     StaticString<42> filename(plane.registration);
     filename += _T(".xcp");
 
-    if (filename != old_filename) {
+    if (copy || filename != old_filename) {
       const auto path = AllocatedPath::Build(old_path.GetParent(),
                                              filename);
 
@@ -275,7 +284,8 @@ PlaneListWidget::EditClicked()
           continue;
       }
 
-      File::Delete(old_path);
+      if (!copy)
+        File::Delete(old_path);
 
       try {
         PlaneGlue::WriteFile(plane, path);
@@ -284,7 +294,7 @@ PlaneListWidget::EditClicked()
         return;
       }
 
-      if (Profile::GetPathIsEqual("PlanePath", old_path)) {
+      if (!copy && Profile::GetPathIsEqual("PlanePath", old_path)) {
         list[index].path = Path(path);
         list[index].name = filename;
         Load(index);
@@ -307,7 +317,7 @@ PlaneListWidget::EditClicked()
 }
 
 inline void
-PlaneListWidget::DeleteClicked()
+PlaneListWidget::DeleteClicked() noexcept
 {
   assert(GetList().GetCursorIndex() < list.size());
 
@@ -339,7 +349,7 @@ PlaneListWidget::OnActivateItem(unsigned i) noexcept
 }
 
 void
-dlgPlanesShowModal()
+dlgPlanesShowModal() noexcept
 {
   TWidgetDialog<PlaneListWidget>
     dialog(WidgetDialog::Full{}, UIGlobals::GetMainWindow(),
